requirements.txt
streamlit
openapi3
pyyaml
requests